Split BuildModelParameters into interface and implementation
